+++ svelte-demo/src/routes/+page.server.ts
@@ -1,0 +1,1 @@
+export const ssr = false;
